writing new querier for the tsdb
